stress_scenario_wizard/requirements.txt
PySide6>=6.7.0
pandas>=2.2.0
numpy>=1.26.0
pyarrow>=17.0.0
sqlalchemy>=2.0.0
pyodbc>=5.1.0
openpyxl>=3.1.2
python-pptx>=1.0.2
reportlab>=4.2.0
plotly>=5.22.0
jinja2>=3.1.4
itsdangerous>=2.2.0
flask>=3.0.0
cryptography>=43.0.0
keyring>=25.0.0
pytest>=8.3.0
pywin32>=306; platform_system == "Windows"
